requirements.txt
aiogram==3.1.1
asyncpg==0.29.0
redis==5.0.1
apscheduler==3.10.4
playwright==1.40.0
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
cryptography==41.0.7
pydantic==2.3.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.1.0
jinja2==3.1.2
SQLAlchemy==2.0.23
psutil==5.9.7
pytest==7.4.3
pytest-asyncio==0.23.2
